% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_list_of_ncfiles}
\alias{check_list_of_ncfiles}
\title{Checks validity for a given list of NetCDF files}
\usage{
check_list_of_ncfiles(nclist)
}
\arguments{
\item{nclist}{A vector of character strings specifying the complete paths to
files.}
}
\value{
NULL if all provided files are valid, otherwise throws an error
listing all the invalid files.
}
\description{
Checks validity for a given list of NetCDF files
}
