% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_df_lon_index}
\alias{get_df_lon_index}
\title{Returns a data.frame defining longitude value and longitude index for a given
NetCDF file}
\usage{
get_df_lon_index(x, lonnam)
}
\arguments{
\item{x}{Either a character string specifying the complete path to a
NetCDF file, or an object of class \code{tidync::tidync}, or a list returned
by \code{get_grid} (containing named elements \code{lon_start},
\code{dlon}, \code{len_ilon},  \code{lat_start}, \code{dlat}, and
\code{len_ilat}).}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}
}
\value{
Tidy tibble containing longidude indices and values of the given file.
}
\description{
Returns a data.frame defining longitude value and longitude index for a given
NetCDF file
}
