% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_file_suffix}
\alias{get_file_suffix}
\title{Creates a string used for file suffix of longitudinal bands}
\usage{
get_file_suffix(ilon, df_lon_index)
}
\arguments{
\item{ilon}{A longitude index}

\item{df_lon_index}{A data frame identifying the longitude value for all
indices, returned by function \code{get_df_lon_index}.}
}
\value{
A character string
}
\description{
Creates a string used for file suffix of longitudinal bands
}
