% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_grid}
\alias{get_grid}
\title{Returns a data.frame defining longitude value and longitude index for a given
NetCDF file}
\usage{
get_grid(ncdf, lonnam, latnam)
}
\arguments{
\item{ncdf}{Either a character string specifying the complete path to a
NetCDF file, or an object of class \code{tidync::tidync}, or a list returned
by \code{get_grid}.}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}

\item{latnam}{The dimension name of latitude in the NetCDF files.}
}
\value{
A list specifying the grid. Contains elements \code{lon_start},
\code{dlon}, \code{len_ilon},  \code{lat_start}, \code{dlat}, and
\code{len_ilat}.
}
\description{
Returns a data.frame defining longitude value and longitude index for a given
NetCDF file
}
