% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_by_chunk.R
\name{get_index_by_chunk}
\alias{get_index_by_chunk}
\title{Returns indices for a chunk}
\usage{
get_index_by_chunk(ichunk, nchunk, nlon)
}
\arguments{
\item{ichunk}{The chunk index, an integer}

\item{nchunk}{The total number of chunks, an integer}

\item{nlon}{The total number of indices (e.g., length of dimension) that are
to be chunked}
}
\value{
A vector of indices for a given chunk
}
\description{
This splits a set of indices into evenly sized chunks and returns the indices
for a given chunk number
}
