% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_longitude_value_indices}
\alias{get_longitude_value_indices}
\title{Returns a data.frame defining longitude value and longitude index for a given
NetCDF file}
\usage{
get_longitude_value_indices(ncdf, lonnam)
}
\arguments{
\item{ncdf}{Either a character string specifying the complete path to a
NetCDF file, or an object of class \code{tidync::tidync}.}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}
}
\value{
Tidy tibble containing the variables 'varnam'.
}
\description{
Returns a data.frame defining longitude value and longitude index for a given
NetCDF file
}
