% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nclist_to_df_byilon}
\alias{nclist_to_df_byilon}
\title{Returns a tidy data frame from a NetCDF file for a longitudinal band}
\usage{
nclist_to_df_byilon(
  nclist,
  ilon,
  outdir,
  fileprefix,
  varnam,
  lonnam,
  latnam,
  timenam,
  timedimnam,
  fgetdate,
  overwrite
)
}
\arguments{
\item{nclist}{A vector of character strings specifying the complete paths to
files.}

\item{ilon}{An integer specifying an individual longitude index for chunking
all processing and writing chunks to separate output files. If provided,
it overrides that the function extracts data for all longitude indices. If
omitted (\code{ilon = NA}), the function returns tidy data for all longitude
indexes.}

\item{outdir}{A character string specifying output directory where data
frames are written using the \code{save} statement. If omitted (defaults to
\code{NA}), a tidy data frame containing all data is returned.}

\item{fileprefix}{A character string specifying the file name prefix.}

\item{varnam}{The variable name(s) for which data is to be read from NetCDF
files.}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}

\item{latnam}{The dimension name of latitude in the NetCDF files.}

\item{timenam}{The name of dimension variable used for timein the NetCDF
files. Defaults to \code{NA}.}

\item{timedimnam}{The name of the dimension (axis) used for time.
Defaults to \code{NA}.}

\item{fgetdate}{A function to derive the date used for the time dimension
based on the file name.}

\item{overwrite}{A logical indicating whether time series files are to be
overwritten.}
}
\value{
not sure
}
\description{
Returns a tidy data frame from a NetCDF file for a longitudinal band
}
