% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nclist_to_df_byfil}
\alias{nclist_to_df_byfil}
\title{Returns a tidy data frame from a NetCDF file, optionally for a longitudinal
band}
\usage{
nclist_to_df_byfil(
  filnam,
  ilon = NA,
  varnam,
  lonnam,
  latnam,
  timenam,
  fgetdate
)
}
\arguments{
\item{filnam}{file name}

\item{ilon}{An integer specifying an individual longitude index for chunking
all processing and writing chunks to separate output files. If provided,
it overrides that the function extracts data for all longitude indices. If
omitted (\code{ilon = NA}), the function returns tidy data for all longitude
indexes.}

\item{varnam}{The variable name(s) for which data is to be read from NetCDF
files.}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}

\item{latnam}{The dimension name of latitude in the NetCDF files.}

\item{timenam}{The name of dimension variable used for timein the NetCDF
files. Defaults to \code{"time"}.}

\item{fgetdate}{A function to derive the date used for the time dimension
based on the file name.}
}
\value{
not sure
}
\description{
Returns a tidy data frame from a NetCDF file, optionally for a longitudinal
band
}
