% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nclist_to_df_byilon}
\alias{nclist_to_df_byilon}
\title{Returns a tidy data.frame from a list of NetCDF file(s),
optionally subsetting a single a longitudinal band}
\usage{
nclist_to_df_byilon(
  nclist,
  ilon,
  outdir,
  fileprefix,
  varnam,
  lonnam,
  latnam,
  timenam,
  fgetdate,
  overwrite
)
}
\arguments{
\item{nclist}{A vector of character strings specifying the complete paths to
files.}

\item{ilon}{An integer specifying an individual longitude index to subset.
If provided, only longitude index 'ilon' is extracted. If omitted
(\code{ilon = NA}), the function returns tidy data for all longitude
indexes.}

\item{outdir}{A character string specifying output directory where data
frames are written using the \code{save} statement. If omitted (defaults to
\code{NA}), a tidy data frame containing all data is returned.}

\item{fileprefix}{A character string specifying the file name prefix.}

\item{varnam}{The variable name(s) for which data is to be read from the
NetCDF files.}

\item{lonnam}{The dimension name of longitude in the NetCDF files.}

\item{latnam}{The dimension name of latitude in the NetCDF files.}

\item{timenam}{The name of dimension variable used for time in the NetCDF
files.}

\item{fgetdate}{A function to derive the date(s) used for the time dimension
based on the file name.}

\item{overwrite}{A logical indicating whether time series files are to be
overwritten.}
}
\value{
Tidy tibble containing the variables 'varnam'.
Tibble contains columns 'lon' (double), 'lat' (double), and a nested
column 'data'.
Column 'data' contains requested variables (probably
as doubles) and might contain a column 'datetime' (as string).
Note that the datetime is defined by package CFtime and can contain
dates such as "2021-02-30 12:00:00", which are valid for 360-day
calendars but not for POSIXt. Because of that these dates need to be
parsed separately.
If \code{!is.na(outdir)}, then an RDS file is generated in
\code{outdir} and column 'data' of return value contains a status
message only instead of data.
}
\description{
Returns a tidy data.frame from a list of NetCDF file(s),
optionally subsetting a single a longitudinal band
}
